Fix issue with GPG PINENTRY prompt interrupting Alfred.
Alfred is sorely lacking in basic features, so it is not possible to
pass around stdin/stdout between scripts without Alfred losing its place
in the Workflow. I.e. when the PINENTRY prompt appeared, and I entered
my credentials, Alfred would then quit the workflow without warning or
error. I have implemented a dodgy workaround. Secrets are stored in a
temporary file created with `mktemp`. Not entirely secure, but good
enough for me, for now.